requirements.txt
py-clob-client>=0.34.0
python-telegram-bot>=20.0
httpx>=0.25.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
